tests/wpt/web-platform-tests/workers/constructors/SharedWorker/unresolvable-url.html

<!doctype html>
<title>resolving broken url</title>
<link rel=help href="http://www.whatwg.org/html/#dom-sharedworker">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws_dom("SyntaxError", () => {
    const worker = new SharedWorker('http://foo bar');
  });
});
</script>
